// smartcoin-frontend/src/app/api/model/categoryResponse.ts
/**
 * Smartcoin API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Una categoría del usuario.
 */
export interface CategoryResponse { 
    id: number;
    name: string;
}